# docs/contracts/1.6/06_OPENAPI.yaml
openapi: 3.1.0
info:
  title: Smart Wardrobe Prototype API
  version: 1.6.0
  description: HLD v1.1 / FSD v1.1 / Backend LLD v1.1 기반. 서버 소유의 REST 계약; Decart·ThinQ
    CLO 외부 실제 API 규격을 주장하지 않음. Bearer 인증과 demo_mode 세션은 프로토타입 정책이며 운영 보안 정책이 아님.
servers:
- url: http://localhost:8000
  description: Local FastAPI prototype
...
security:
- BearerAuth: []
paths:
...
  /api/v1/cards/drafts:
    post:
      tags:
      - Cards
      summary: 코디카드 데이터·템플릿 구성
      operationId: post_api_v1_cards_drafts
      x-fsd-ids:
      - CARD-001
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardPrepareRequest'
      responses:
        '201':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDraft'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/card-render-jobs:
    post:
      tags:
      - Cards
      summary: 카드 이미지 비동기 렌더링
      operationId: post_api_v1_card_render_jobs
      x-fsd-ids:
      - CARD-002
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRenderRequest'
      responses:
        '202':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJob'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cards:
    get:
      tags:
      - Cards
      summary: 저장된 코디카드 조회
      operationId: get_api_v1_cards
      x-fsd-ids:
      - CARD-003
      parameters:
      - *id001
      - *id004
      - *id005
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDraftList'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cards/{card_id}/save:
    post:
      tags:
      - Cards
      summary: 카드 저장 및 공유 정책 설정
      operationId: post_api_v1_cards_card_id_save
      x-fsd-ids:
      - CARD-003
      parameters:
      - name: card_id
        in: path
        required: true
        schema: *id002
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardSaveRequest'
      responses:
        '201':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDraft'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cards/{card_id}:
    get:
      tags:
      - Cards
      summary: 카드 상세·재사용 정보
      operationId: get_api_v1_cards_card_id
      x-fsd-ids:
      - CARD-003
      parameters:
      - name: card_id
        in: path
        required: true
        schema: *id002
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDraft'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/card-shares/{token}:
    get:
      tags:
      - Cards
      summary: Explicit expiring card image share
      operationId: get_api_v1_card_shares_token
      x-fsd-ids:
      - CARD-003
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Private image shared by explicit bearer link
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 서버가 발급한 사용자 범위 세션 토큰. 요청 member_id/household_id를 신뢰하지 않고 인증 주체와
        비교.
  schemas:
    Error:
      type: object
      properties:
        code: *id003
        message: *id003
        request_id: *id003
        details:
          type: object
          additionalProperties: true
      required:
      - code
      - message
      - request_id
      additionalProperties: false
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/Error'
      required:
      - error
      additionalProperties: false
    Page:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      required:
      - limit
      - offset
      - total
      additionalProperties: false
    AcceptedJob:
      additionalProperties: false
      properties:
        job_id:
          format: uuid
          type: string
        status:
          enum:
          - QUEUED
          - RUNNING
          - SUCCEEDED
          - READY
          - FAILED
          - TIMED_OUT
          - CANCELLED
          type: string
        status_url:
          type: string
      required:
      - job_id
      - status
      - status_url
      type: object
    Asset:
      type: object
      properties:
        asset_id: *id002
        asset_key: *id003
        content_type: *id003
        read_url:
          anyOf:
          - type: string
          - type: 'null'
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
      required:
      - asset_id
      - asset_key
      - content_type
      additionalProperties: false
...
    CardTemplateData:
      additionalProperties: false
      properties:
        title:
          type: string
        style_tag:
          type: string
        date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
        weather_label:
          anyOf:
          - type: string
          - type: 'null'
        items:
          items:
            $ref: '#/components/schemas/CardItem'
          type: array
        template_version:
          const: minimal-v1
          type: string
      required:
      - title
      - style_tag
      - items
      type: object
    CardPrepareRequest:
      additionalProperties: false
      properties:
        outfit_id:
          format: uuid
          type: string
        context_snapshot_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
        template_id:
          const: minimal-v1
          type: string
      required:
      - outfit_id
      - template_id
      type: object
    CardDraft:
      additionalProperties: false
      properties:
        id:
          format: uuid
          type: string
        outfit_id:
          format: uuid
          type: string
        template_id:
          type: string
        data:
          $ref: '#/components/schemas/CardTemplateData'
        status:
          enum:
          - DRAFT
          - QUEUED
          - RUNNING
          - READY
          - FAILED
          - SAVED
          type: string
        image_asset:
          anyOf:
          - $ref: '#/components/schemas/Asset'
          - type: 'null'
        is_saved:
          type: boolean
        saved_title:
          anyOf:
          - type: string
          - type: 'null'
        share_url:
          anyOf:
          - type: string
          - type: 'null'
        share_expires_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        shared_fields:
          items:
            type: string
          type: array
        linked_outfit_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
      required:
      - id
      - outfit_id
      - template_id
      - data
      - status
      type: object
    CardRenderRequest:
      additionalProperties: false
      properties:
        card_id:
          format: uuid
          type: string
        output_format:
          enum:
          - PNG
          - WEBP
          type: string
        width:
          const: 1080
          type: integer
        height:
          const: 1350
          type: integer
      required:
      - card_id
      type: object
    CardSaveRequest:
      additionalProperties: false
      properties:
        title:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
        visibility:
          enum:
          - PRIVATE
          - SHAREABLE
          type: string
        reuse_outfit:
          type: boolean
      required:
      - visibility
      type: object
...
    CardDraftList:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: '#/components/schemas/CardDraft'
          type: array
        page:
          $ref: '#/components/schemas/Page'
      required:
      - items
      - page
      type: object
...
    CardItem:
      additionalProperties: false
      properties:
        garment_id:
          format: uuid
          type: string
        slot:
          type: string
        position:
          type: integer
        image_asset_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
      required:
      - garment_id
      - slot
      - position
      type: object
...